+++ android/art/runtime/base/locks.h
@@ -23,8 +23,8 @@ namespace art {
 
 class LockLevel {
 public:
-    static uint8_t kMonitorLock;
-    static uint8_t kLockLevelCount;
+    static uint32_t kMonitorLock;
+    static uint32_t kLockLevelCount;
     static void Init();
     static void Init23();
     static void Init24();
